Lowered RasterType and now it gives a speedup, TODO cleanup

+++ src/types.rs
@@ -2,7 +2,7 @@ use nalgebra::{Const, Dyn, OMatrix, OVector};
 
 pub type FloatType = f64;
 pub type IndexType = u16;
-pub type RasterType = i64;
+pub type RasterType = i16;
 
 pub type FloatMatrixType<const NCOLS: usize> = OMatrix<FloatType, Dyn, Const<NCOLS>>;
 pub type IndexVectorType = OVector<IndexType, Dyn>;
